Solution to Input Two Floats and Print Their Product

+++ Input Two Floats and Print Their Product@@6788d5012db25efec946f93e/code.c
@@ -1,9 +1,9 @@
 #include<stdio.h>
 int main()
 {
-    float x,y;
-    scanf("%f %f",&x,&y);
-    float z=x*y;
-    printf("Product: %f",z);
+    double x,y;
+    scanf("%lf %lf",&x,&y);
+    double z=x*y;
+    printf("Product: %lf",z);
     return 0;
 }
